requirements.txt
numpy # Don't remove
Pillow # Don't remove
mmcv-full